+++ TabBar/index.js
@@ -10,6 +10,8 @@ import {
 
 import transPxToDp from '../util/transPxToDp';
 
+const NOOP = () => {};
+
 const styles = StyleSheet.create({
   all: {
     backgroundColor: '#fff',
@@ -21,25 +23,8 @@ const styles = StyleSheet.create({
 });
 
 class TabBar extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      activedItemId: props.items[0].id,
-    };
-  }
-
   onPress() {
-    if (this.context.state.activedItemId !== this.id) {
-      const res = this.context.props.onChange(this.id);
-
-      if (res === false) {
-        return;
-      }
-
-      this.context.setState({
-        activedItemId: this.id,
-      });
-    }
+    this.context.props.onPress(this.id);
   }
 
   render() {
@@ -55,7 +40,7 @@ class TabBar extends Component {
               activeOpacity={this.props.activeOpacity}
             >
               {
-                this.state.activedItemId === item.id ?
+                this.props.activeId === item.id ?
                 item.activedComponent : item.defaultComponent
               }
             </TouchableOpacity>
@@ -67,6 +52,8 @@ class TabBar extends Component {
 }
 
 TabBar.propTypes = {
+  // 激活项的 Id，必须是 items 项的 id 属性
+  activeId: PropTypes.string,
   // 自定义样式
   style: View.propTypes.style,
   // tab 项
@@ -80,15 +67,15 @@ TabBar.propTypes = {
   })),
   // tab 项点击时透明度
   activeOpacity: PropTypes.number,
-  // 改变激活项时的回调，激活项的 id 会作为参数传入。当该函数返回 false 时，可以阻止当前切换操作。
-  onChange: PropTypes.func,
+  // 改变激活项时的回调，激活项的 id 会作为参数传入。
+  onPress: PropTypes.func,
 };
 TabBar.defaultProps = {
+  activeId: '',
   style: null,
   items: [{}],
   activeOpacity: 1,
-  onChange: () => {},
+  onPress: NOOP,
 };
-
 
 export default TabBar;